Fixed comments
Cleaned up comments - lots of spelling errors fixed.

+++ src/Zanshin_BME680.cpp
@@ -369,7 +369,7 @@ const uint32_t lookupTable2[16]  = {
 	var1      = ((32768 + var1) * (int32_t)_P1) >> 15;
 	_Pressure = 1048576 - adc_pres;
 	_Pressure = (int32_t)((_Pressure - (var2 >> 12)) * ((uint32_t)3125));
-	var4      = (1 << 31);
+	var4      = ((int32_t)1 << 31);
 	if (_Pressure >= var4)
     _Pressure = ((_Pressure / (uint32_t)var1) << 1);
 	else
